[Add] SharePostsPage.jsxのモックを作成
