+++ front/src/templates/index.js
@@ -4,4 +4,5 @@ export {default as PostEdit} from './PostEdit'
 export {default as ReadingPostsDetail} from './ReadingPostsDetail'
 export {default as RegisteredPostsDetail} from './RegisteredPostsDetail'
 export {default as RegisteredPostsList} from './RegisteredPostsList'
+export {default as SharePostsPage} from './SharePostsPage'
 export {default as TopPage} from './TopPage'
